# src_backup/app/api/hr/openapi.yaml
openapi: 3.0.0
info:
  title: HR & Asset Management API
  description: API documentation for the Human Resources & Asset Management module of the Hospital Management System
  version: 1.0.0
  contact:
    name: HMS Support
    email: support@hms.example.com
servers:
  - url: /api
    description: API base URL
...
paths:
...
  /hr/payroll/periods:
    get:
      summary: List payroll periods
      description: Returns a paginated list of payroll periods
      tags:
        - Payroll
      parameters:
        - name: skip
          in: query
          description: Number of records to skip for pagination
          schema:
            type: integer
            default: 0
        - name: take
          in: query
          description: Number of records to return per page
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [DRAFT, PROCESSING, APPROVED, PAID]
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  periods:
                    type: array
                    items:
                      $ref: "#/components/schemas/PayrollPeriod"
                  total:
                    type: integer
                  skip:
                    type: integer
                  take:
                    type: integer
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create payroll period
      description: Creates a new payroll period
      tags:
        - Payroll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollPeriodInput"
      responses:
        "200":
          description: Payroll period created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollPeriod"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /hr/biomedical/{id}/calibration:
    get:
      summary: Get calibration records
      description: Returns calibration records for a specific biomedical equipment
      tags:
        - Biomedical
      parameters:
        - name: id
          in: path
          required: true
          description: Biomedical equipment ID
          schema:
            type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalibrationRecord"
        "404":
          description: Biomedical equipment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Record calibration
      description: Records calibration for a specific biomedical equipment
      tags:
        - Biomedical
      parameters:
        - name: id
          in: path
          required: true
          description: Biomedical equipment ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalibrationInput"
      responses:
        "200":
          description: Calibration recorded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalibrationRecord"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "404":
          description: Biomedical equipment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    PayrollPeriod:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        paymentDate:
          type: string
          format: date
        status:
          type: string
          enum: [DRAFT, PROCESSING, APPROVED, PAID]
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        approvedBy:
          type: string
        approvedAt:
          type: string
          format: date-time
        paidBy:
          type: string
        paidAt:
          type: string
          format: date-time
    PayrollPeriodInput:
      type: object
      required:
        - name
        - startDate
        - endDate
        - paymentDate
      properties:
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        paymentDate:
          type: string
          format: date
        status:
          type: string
          enum: [DRAFT]
          default: DRAFT
...
    CalibrationRecord:
      type: object
      properties:
        id:
          type: string
        biomedicalEquipmentId:
          type: string
        date:
          type: string
          format: date
        performedBy:
          type: string
        result:
          type: string
          enum: [PASS, FAIL, ADJUSTED]
        notes:
          type: string
        attachments:
          type: array
          items:
            type: string
    CalibrationInput:
      type: object
      required:
        - date
        - result
      properties:
        date:
          type: string
          format: date
        performedBy:
          type: string
        result:
          type: string
          enum: [PASS, FAIL, ADJUSTED]
        notes:
          type: string
        nextCalibrationDate:
          type: string
          format: date
        attachments:
          type: array
          items:
            type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
    ValidationError:
      type: object
      properties:
        error:
          type: string
        details:
          type: object